fixed the advanced view to display the proper information and changed the increase, reduce, dispose, and report missing front end logic

+++ WebContent/js/InventoryItemManagement/ItemDispose.js
@@ -7,6 +7,7 @@ let _disposeItemDepartment = '';
 let _disposeItemCategory = '';
 let _disposeItemModel = '';
 let _disposeItemLocation = '';
+let _disposeOnAdvancedView = false;
 
 let _disposeRowId = 0;
 
@@ -15,29 +16,56 @@ const disposeForm = document.getElementById('disposeForm');
 
 // function to handle the response data
 const handleDisposeResponse = (response) => {
-	deleteItem(_disposeRowId, _disposeItemName);
-	showSuccessMessage('Item Successfully Disposed')
-	$('#disposeModal').modal('hide');
-  console.log(response);
+  if (response.deleted && !_disposeOnAdvancedView) {
+	  showErrorMessageOnIncreaseModal('Item Deleted by Another Member')
+	  deleteItem(_disposeRowId, _disposeItemName);
+	  setTimeout(() => {
+	  	$('#disposeModal').modal('hide');
+	  }, 3000)
+  }
+  
+  if (response.deleted && _disposeOnAdvancedView) {
+	  showErrorMessageOnIncreaseModal('Item Deleted by Another Member');
+	  deleteItem(_disposeItemID, _disposeItemName);
+	  deleteItemOnAdvancedView()
+	  setTimeout(() => {
+	  	$('#disposeModal').modal('hide');
+	  }, 3000)
+  }
+  
+  if(!response.modifiedByOtherMember && !response.deleted && !_disposeOnAdvancedView) {
+	  deleteItem(_disposeRowId, _disposeItemName);
+	  showSuccessMessage('Successfully Disposed Item');
+	  $('#disposeModal').modal('hide');
+  }
+  
+  if(!response.modifiedByOtherMember && !response.deleted && _disposeOnAdvancedView) {
+	  deleteItem(_disposeRowId, _disposeItemName);
+	  deleteItemOnAdvancedView()
+	  showSuccessMessage('Successfully Disposed Item');
+	  $('#disposeModal').modal('hide');
+  }
+	
 };
 
 // handler for sending and recieving data from backend
 const disposeItemInDB = () => {
   const servletParameters = {
-    'item-id': _disposeItemID,
-    'item-name': _disposeItemName,
-    'item-quantity': _disposeItemQuantity,
-    'item-price': _disposeItemPrice,
-    'item-department': _disposeItemDepartment,
-    'item-category': _disposeItemCategory,
-    'item-model': _disposeItemModel,
-    'item-location': _disposeItemLocation,
-    'member-id': LOGGED_ON_MEMBER_ID,
+    "item-id": _disposeItemID,
+    "item-name": _disposeItemName,
+    "item-quantity": _disposeItemQuantity,
+    "item-price": _disposeItemPrice,
+    "item-department": _disposeItemDepartment,
+    "item-category": _disposeItemCategory,
+    "item-model": _disposeItemModel,
+    "item-location": _disposeItemLocation,
+    "member-id": LOGGED_ON_MEMBER_ID,
+    "update_type": "dispose"
   };
   $.ajax({
     url: 'ItemDispose',
     dataType: 'text',
-    type: 'DELETE',
+    type: 'POST',
     data: servletParameters,
     success: function (data) {
       let response = JSON.parse(data);
@@ -73,6 +101,7 @@ $('#disposeModal').on('show.bs.modal', function (e) {
   _disposeRowId = $(e.relatedTarget.parentElement.parentElement).data(
     'rowNumber'
   );
+  _disposeOnAdvancedView = $(e.relatedTarget).data('advanced');
 
   document.getElementById('itemIdDisposeModal').innerText = _disposeItemID;
   document.getElementById('itemNameDisposeModal').innerText = _disposeItemName;
